Fix histograms erroneous data ( issue #1 )
Commit ff3218a304a5f421c8f45167d93572b0d5b95726 introduced a bug that
provided the same data to all histograms, rendering all but one of them
useless. The fix has been to promote to the arglist the knowledge about
internal layout of Data.

This should be improved in future commits by using a method to create
the list of data values for the histogram (and offer those data points
to other functions as well). It would remove the need to know about
internal layout details and decouple Data from CLI code. Issue #2 will
track progress on this enhancement.